input shows as invalid on invalid inputs
