+++ src/main/java/org/voyager/controller/BrowseController.java
@@ -1,6 +1,7 @@
 package org.voyager.controller;
 
 import jakarta.validation.Valid;
+import lombok.NonNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
