+++ src/main/resources/templates/fragments/routes.html
@@ -38,15 +38,11 @@
             </div>
             <div id="select-start" class="mt-4 d-grid">
                 <div class="input-group row" style="margin-left:.0em;">
-                    <select class="form-select" style="max-width:200px" disabled>
+                    <select class="form-select" style="max-width:170px" disabled>
                         <option selected disabled>Select Start</option>
                     </select>
                     <input class="form-control" disabled>
                 </div>
-            </div>
-            <div th:replace="~{fragments/form :: airport-form-initial(true)}"></div>
-            <div id="non-delta-start"
-                 hx-swap-oob="outerHTML:#non-delta-start">
             </div>
         </div>
     </div>
@@ -54,7 +50,7 @@
 <!--Select Start Filter and Options-->
 <div th:fragment="selected-start-locations(selection,filterList,optionList)"
     class="input-group row" style="margin-left:.0em;">
-    <select class="form-select" style="max-width:200px">
+    <select class="form-select" style="max-width:170px">
         <option th:each="filter,iter : ${filterList}" th:value="${filter.value}" th:text="${filter.display}"></option>
     </select>
     <datalist th:if="${selection == 'AIRPORT'}" id="select-start-options">
@@ -63,7 +59,7 @@
 <!-- TODO: only allow add to map if matches airport filter -->
     <input th:if="${selection == 'AIRPORT'}" class="form-control"
            placeholder="Enter Airport" list="select-start-options"
-           oninput="addAirportToMap(this,true)"
+           oninput="debouncedAddAirportToMap(this,true)"
     >
     <select th:if="${selection == 'LOCATION'}" class="form-control">
         <option th:each="option,iter : ${optionList}" th:value="${option.value}" th:text="${option.display}"></option>
